+++ src/Exception/TransformationException.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: oguzu
+ * Date: 1-4-2017
+ * Time: 14:24
+ */
+
+namespace Pbxg33k\MusicInfo\Exception;
+
+
+class TransformationException extends \Exception
+{
+
+}